Use separate request and response schema.
